# lib/api-zod/src/generated/types/batchVerification.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BatchVerification {
  authentic: boolean;
  message: string;
  productName: string;
  batchNumber: string;
  expiryDate: string;
  sellerVerified: boolean;
}