# lib/api-zod/src/generated/types/batchVerificationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BatchVerificationInput {
  /** @minLength 1 */
  qrPayload: string;
}