# lib/api-zod/src/generated/types/cartItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CartItem {
  productId: string;
  batchId: string;
  /** @minimum 1 */
  quantity: number;
}