# lib/api-zod/src/generated/types/diagnosis.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Differential } from './differential';

export interface Diagnosis {
  id: string;
  cropName: string;
  condition: string;
  category: string;
  confidence: number;
  severity: string;
  evidence: string[];
  differentials: Differential[];
  suggestedObservations: string[];
  requiresOfficerReview: boolean;
  provenance: string;
  safetyNote: string;
  imageDataUrl?: string;
}