# lib/api-zod/src/generated/types/escalation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Escalation {
  id: string;
  diagnosisId: string;
  cropName: string;
  condition: string;
  confidence: number;
  severity: string;
  status: string;
  createdAt: string;
  reason: string;
  /** @nullable */
  officerNotes?: string | null;
}