# lib/api-zod/src/generated/types/escalationReviewInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EscalationReviewInputDecision } from './escalationReviewInputDecision';

export interface EscalationReviewInput {
  decision: EscalationReviewInputDecision;
  /** @minLength 3 */
  notes: string;
  overrideCondition?: string;
}